feat(mcp): add inbound MCP smoke tests and runbook
Keep `mcp server` on a thin stdio entry so host-spawned children never boot SQLite or native addons, and cover the handshake, idle/EOF, and error paths with a child-process e2e plus an operations runbook.
